fix: fix value_money to return double

+++ src/main/java/dev/pantanal/b3/krpv/acao_social/modulos/Investment/InvestmentEntity.java
@@ -44,7 +44,7 @@ public class InvestmentEntity {
 
     @Column(nullable = false)
     //@NotBlank
-    private long value_money;
+    private Double value_money;
 
     @Column(nullable = false)
     //@NotBlank
